requirements.txt
accelerate==0.33.0
datasets==2.16.0
gdown==5.2.0
huggingface_hub==0.24.2
numpy==1.24.3
opencv_python==4.10.0.84
pandas==1.3.5
Pillow==10.1.0
pycocoevalcap==1.2
pycocotools==2.0.7
pynvml==11.5.3
scikit_learn==1.2.2
scipy==1.14.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.28.1
